Rename Dates to Dates/TrainingDates, it's used as $trainingDates anyway
