+++ site/app/Articles/Blog/BlogPostPreview.php
@@ -4,7 +4,7 @@ declare(strict_types = 1);
 namespace MichalSpacekCz\Articles\Blog;
 
 use MichalSpacekCz\Formatter\TexyFormatter;
-use MichalSpacekCz\Training\Dates;
+use MichalSpacekCz\Training\Dates\TrainingDates;
 use Nette\Bridges\ApplicationLatte\DefaultTemplate;
 use Spaze\ContentSecurityPolicy\CspConfig;
 
@@ -14,7 +14,7 @@ class BlogPostPreview
 	public function __construct(
 		private readonly TexyFormatter $texyFormatter,
 		private readonly BlogPosts $blogPosts,
-		private readonly Dates $trainingDates,
+		private readonly TrainingDates $trainingDates,
 		private readonly CspConfig $contentSecurityPolicy,
 	) {
 	}
