+++ site/app/Training/CompanyTrainings.php
@@ -5,6 +5,7 @@ namespace MichalSpacekCz\Training;
 
 use Contributte\Translation\Translator;
 use MichalSpacekCz\Formatter\TexyFormatter;
+use MichalSpacekCz\Training\Dates\TrainingDates;
 use Nette\Database\Explorer;
 use Nette\Database\Row;
 
@@ -14,7 +15,7 @@ class CompanyTrainings
 	public function __construct(
 		private readonly Explorer $database,
 		private readonly TexyFormatter $texyFormatter,
-		private readonly Dates $trainingDates,
+		private readonly TrainingDates $trainingDates,
 		private readonly Translator $translator,
 	) {
 	}
