+++ site/config/services.neon
@@ -78,7 +78,7 @@ services:
 	- MichalSpacekCz\Tls\CertificatesApiClient
 	- MichalSpacekCz\Training\Applications(emailEncryption: @emailEncryption)
 	- MichalSpacekCz\Training\CompanyTrainings
-	- MichalSpacekCz\Training\Dates
+	- MichalSpacekCz\Training\Dates\TrainingDates
 	- MichalSpacekCz\Training\Dates\TrainingDatesFormValidator
 	- MichalSpacekCz\Training\Files\TrainingFileFactory
 	- MichalSpacekCz\Training\Files\TrainingFiles
